Use module dict to get device array capsule pointer
When `import_devicearray()` is called, we're partway through importing
`numba.cuda`. Therefore, the `PyCapsule_Import()` fails because it tries
to access packages under `numba.cuda` during its initialization, which
then fails due to this circularity.

In order to work around this, we can get the `_DEVICEARRAY_API`
attribute of the `_devicearray` module directly from its module dict.

This was not a problem in upstream because `_devicearray` was not in the
`numba.cuda` package.

NOTE: This fix is not entirely complete - it is not good at cleaning up
in the presence of errors occurring - this will be resolved in a later
commit when the debug prints have been tidied up.