Refactor cuda_paths.py, support CUDA 13 conda packages better (#464)
Fixes #452

Refactors a large chunk of `cuda_paths.py` in order to properly support
all of the package types that we support. Removes some outdated
functionality for things like when the NVIDIA channel conda packages
diverged from conda-forge, debian packages, etc.

It was extremely difficult to follow the existing logic of when
directory paths vs file paths were returned from the `get_cuda_paths`
function and it seemed to differ based on the source of the package, so
everything should now be unified to either always return a file path
(suffixed by `_path`) or a directory path (suffixed by `_dir`).