Add `compile_all` API (#484)
This PR adds `compile_all` API. Instead of only returning the compiled
kernel, it also compiles cuda source files for its external linking
files. In the case that external files are not compilable (such as
cubin, ptx, LTOIR etc.), they will be passed through into the return
list.

Testing of `compile_all` completely is tough due to combinatorial
complexity from the number of arguments. Since it's mainly a passthrough
of `compile` with additional nvrtc pass on external linking files, I
reused all existing tests in `TestCompile` to make sure the API is
consistent with previous APIs.

The default arguments of `compile_all` is the same as `compile`.

When `lineinfo` or `debug` is set to true, the nvrtc flag for either
options are set accordingly for external files.

---------

Co-authored-by: Michael Wang <isVoid@users.noreply.github.com>
Co-authored-by: Graham Markall <535640+gmarkall@users.noreply.github.com>