Improve debug value range coverage (#461)
The current kernel debug support tracks debug value of variables by
inserting dbg.value immediate after a store instruction to a user
variable. However, in some specific scenario the debug value info could
be missing range, see details in nvbug#5321343, when a basic block has a
load instruction loading from a user variable and saving to a temparary,
then store the temporary to an internal variable without any other
assignment instruction saving value to that user variable in the whole
basic block. The back end libnvvm can not fully track the variable value
in this scenario.

This PR adds the enhanced handling of debug value, detecting the not yet
covered user variable and inserting the dbg.value for the missing range
per basic block.

A test is also added to guard this change.