Ensure NUMBA can manipulate memory from CUDA graphs before the graph is launched (#437)
When using CUDA graphs, memory nodes correspond to memory that is
reserved in virtual memory, but only allocated when the graph is
launched. This mean we cannot check whether a pointer is a valid device
address by the time we create memory objects that we will pass to
kernels.

We now use the pointer from CUDA Array Interface objects directly instead of
getting the `CU_POINTER_ATTRIBUTE_DEVICE_POINTER` attribute of them, so that
objects backed by memory nodes in CUDA graphs can be used.

This is expected to be safe as all CUDA processes are 64-bit and with devices of
Compute Capability 2.0 or greater, so a Unified Virtual Address Space will be in
use. These requirements are documented in:

https://docs.nvidia.com/cuda/cuda-c-programming-guide/index.html?#unified-virtual-address-space

---------

Co-authored-by: Graham Markall <gmarkall@nvidia.com>