Don't use `MemoryLeakMixin` for tests that don't use NRT
Some recent vendorings resulted in the `MemoryLeakMixin` class being a
base for some tests. The side effect of this is that it now enables NRT
stats counting on-device. There seem to be some cleanup issues with NRT
stats counting (it ought to be robust when used with tests in general,
but as #517 notes, it seems not to be).

In order to reduce the incidence of spurious failures from this
(especially when testing on a fast machine), we disable the
`MemoryLeakMixin` for tests that aren't using NRT (which is still
experimental for CUDA anyway).