pyproject.toml: add search path for Pyrefly
This allows pyrefly to locate `numba.cuda` code within the `numba_cuda`
package instead of in the installed Numba package. With this change,
using Pyrefly as an LSP behaves "as expected" - i.e. suggestions / info
based on code in `numba_cuda` are provided instead of those from the
legacy upstream CUDA target.