Draft: Vendor in the IR module (#439)
The IR module is near the root of the import graph of numba-cuda, so it
has many dependencies.

The plan-of-record as discussed offline with @gmarkall is to vendor this
module piecemeal; modules in numba-cuda will continue to use numba until
we have vendored in additional modules (such as types and typing).