perf: remove context threading in various pointer abstractions (#536)
This PR removes the `context` member of the various pointer abstractions
in the codebase.

None of these `context`s were ever being used, except as arguments to
constructors of these types.

The main implication is that `from_cuda_array_interface` no longer
needs either `require_context` nor `current_context()` (since that
seemed to be the reason for having the decorator).

Over thousands of calls to `numba.cuda.as_cuda_array()`, this
repeated context acquisition is expensive.

xref: https://github.com/numba/numba/issues/9835

This is not the only overhead for that function that we can address, but
this is a set of self-contained changes.

I will open separate PRs for the others.