test: speed up ipc tests by ~6.5x (#527)
This PR speeds up multiprocessing-based IPC tests by using the
concurrent futures interface.

This is faster because processes are reused from a pool as opposed to
creating all the machinery
to run a thing in new process in every single test or `subTest`
invocation.

While this is less isolated, I would argue that it's a much more common
use case to use pools and/or
the concurrent futures interface (it's much less noisy, for example).