perf: cache dimension computations (#542)
This PR adds some `functools.cache` decorations to a few functions for
computing dimensions, which are repeatedly computed for the same inputs
many times.

This does trade some space to reduce duration of conversion. That space will
scale with the number of unique combinations of shape, strides, dtypes, and
memory ordering.

If someone has a lot of arrays that have different shapes,
strides, dtypes, and memory ordering, the cache can potentially use a
lot of memory.

I'm not sure that's a likely scenario, but it's worth pointing out as a potential issue.