Vendor in types and datamodel for CUDA-specific changes (#533)
This is a large PR that vendors in `types` and `datamodel` from Numba
for future CUDA-specific customizations.

This PR includes redirection to the Numba types and datamodel if they
are present; handling of Numba-CUDA types and datamodel in
conjunction with the Numba types and datamodel is left for future work.

---------

Co-authored-by: Graham Markall <gmarkall@nvidia.com>