Fix registration with Numba, vendor MakeFunctionToJITFunction tests
PR #555 removed dependencies on the target extension, but it also
removed registration in Numba's target registry. This prevents handling
of closures, because closure handling uses a process that is equivalent
to wrapping closure functions in a `@jit` decorator, where the
appropriate `@jit` decorator for the target is looked up in the jit
registry.

This commit restores the registration when Numba is present, and vendors
/ ports `test_make_function_to_jit_function`. This ensures that the
target registration continues to work, and also exercises the
`MakeFunctionToJITFunction` pass, which previously had no effect on any
code in the test suite.

In order to preserve the form of the original tests, the original test
code has the `@njit` decorators replaced with a wrapper function that
generates an appropriate kernel, temporary storage, and launch code.

Some changes had to be made to avoid refcounting, where arrays would
have been used. These have been replaced by (in different cases):

- Summation of all the values that would have been array elements, or
- Carefully constructing tuples to return and then packing / unpacking
  them at function boundaries.