[Refactor][NFC][Cleanups] Update imports to upstream numba to use the numba.cuda modules (#561)
This PR updates the vast majority of imports to upstream numba modules,
that had already been vendored in for future CUDA-specific changes.
After this PR, there should be exactly 3 (not strictly necessary)
upstream numba imports remaining - one in testing vectorization for in
test_math, one in random.py, and the global compiler_lock.

It seems from the existing comments that removing the CPU jit in
random.py would result in considerable performance drops, so I've just
added a guard for numba being available.

Some lowering / overload implementations that depend on
numba.core.misc.{mergesort,quicksort} have been removed. These
are presently not regarded as supported, and are unlikely to be
performant implementations in a CUDA-specific context.

There are many tests that rely on CPU jitting, these tests are now
guarded to not run in environments where the numba package is not
available. This is done using a global importlib check in numba-cuda's
`__init__.py`. This cannot be moved into numba.cuda.config since the
upstream numba doesn't have this flag, and we intend on forwarding calls
to upstream numba when it is available. All previously sporadic guards
for this *should* have been updated to use this global flag now.

Support for prange, pndindex, and stencils has been removed, likewise
for typed.Dict, and typed.List.

 Also, this removes all mention of the legacy type
system flag since we implicitly assume the legacy type system in place.