Fix test_native_cast
Some NVVM versions will generate a store using `b<N>` instead of `f<N>`.
This is legitimate, as the argument and return types in this test are
inferred to be the same (i.e. casting `float32` to `float32`, and
`float16` to `float16`); the behaviour of `float()` on a Numba floating
point type is to preserve its bitwidth.