Fix freezing in of constant arrays with negative strides (#589)
This PR fixes a bug when freezing certain views of numpy arrays into
kernels. The current implementation loops through the bytes of the
source array in viewed order but maintains a potentially negative stride
without also in that case bumping the pointer to the end of the array.
This PR changes it so that the physical array we put in constant memory
is just the contiguous logical equivalent of the original array with a
positive stride. This should be simpler than actually jumping through
the pointer arithmetic given we have to make a copy of the view anyways,
so the inherent "viewness" is already going to be lost when moving the
data to device.