Add more thirdparty tests (#586)
This adds tests of additional third party libraries and fixes up issues
that were found during testing them. Notes on the changes:

* The latest release, 0.6.0, of nvmath-python is used, and its device
API tests are run.
* nvmath-python patches `types.number_domain`, which was not handled by
our vendoring. The changes in `cudaimpl.py` resolve this issue.
* nvmath-python also creates extensions that can lead to a Numba-CUDA
`Signature` class leaking into Numba, which checks for an instance of a
Numba `Signature` only. The change in `templates.py` mitigates this by
always constructing core `Signature` instances, if Numba is present.
* I tested the code with my ["Extending Numba CUDA"
tutorials](https://github.com/gmarkall/extending-numba-cuda), and
noticed that the `Dim3` and `GridGroup` were not in our `types` package,
so I've added these back. I haven't added these examples as part of our
CI as there would be some extra work involved in converting them to
tests, but I may add them in a future PR.
* Awkward 2.8.10 is used for tests, as outlined in
https://github.com/scikit-hep/awkward/discussions/3587
* The nvmath-python tests take 45 minutes so I've set them to only run
on pushes to `main` rather than slowing down CI for PR testing. This is
a little less than ideal, but better than increasing the iteration time
for testing in CI. This is done in a separate workflow, `ci.yaml`
* It may be possible to run a reduced set of nvmath-python tests for
every PR, but I need to work with the nvmath-python team to understand
how to do that without sacrificing too much coverage.

---------

Co-authored-by: Marcus D. Hanwell <mhanwell@nvidia.com>