tailor for numba-cuda
We use a separate job to build test artifacts:

Ideally we want to overlay two CTKs in the same test job, which was proven possible
in cuda-python. However, it limits to using CTK headers only; if we need to manipulate
PATH (which is necessary for finding both EXE and DLL on Windows), this is not trivial
because GHA does not allow us to manipulate GITHUB_PATH easily. The WAR is to parametrize
different CTK versions, and for that we need a separate job...