Attempt to fix Awkward Array 3rd party tests (#607)
A couple of Awkward Array tests cover functionality that is experimental
and/or unused in CUDA, as discussed in
https://github.com/scikit-hep/awkward/discussions/3587.

These tests can fail intermittently, causing issues for CI. This PR
patches out the generation and execution of these tests.

We should also not be constraining the version of cuda-core that is
installed alongside numba-cuda. The Awkward Array CI test script did
this, but so did the other thirdparty tests and the coverage test. This
pinning is now removed.