Add DWARF address class support for shared memory arrays (#594)
This change adds "dwarfAddressSpace" attribute to debug metadata for
CUDA shared memory pointers, enabling debuggers to correctly identify
memory location of variables.

I choose to add address space tracking in the lowering phase, rather
than modifying the underlying typing infrastructure (ArrayModel,
PointerModel) due to the following reasons:
1) There is an onging effort decoupling from Numba's typing system, but
the default behavior is still redirect to Numba;
2) There is a WIP
[PR#236](https://github.com/NVIDIA/numba-cuda/pull/236) introducing
CUDAArray type and implementation with addresspace information.

When either of the above is completed, there will be a cleaner approach
to update this patch.

So in this change,
1) Add detection in CUDALower Numba ir.Call to find cuda.shared.array()
call; set flag for the subsequent storevar() to record the name /
addrespace mapping; later reference the address space map when emitting
debug info.
2) A mapping from NVVM address space to DWARF address class is added in
order to emit the "dwarfAddressSpace" to the DIDerivedType for pointer
member "data" from the CUDA array descriptor.
3) A new test is added to make sure shared array and regular local array
get distinguished.

This fixes nvbug#5643016.

---------

Co-authored-by: Graham Markall <gmarkall@nvidia.com>