Fix #624: Accept Numba IR nodes in all parts of Numba-CUDA IR
There were some specific isinstance checks in `ir.py` that were modified
to accept Numba IR nodes that could have been generated when
transformations in Numba or other areas produce them. These specific
locations appear to have been insufficient, leading to #624.

In order to ensure that Numba IR nodes are accepted in general, all
isinstance checks are updated to accept either the Numba-CUDA type or
the Numba type.

Rather than implementing a lot of switching in the implementations, the
sets of acceptable IR nodes are defined in the module global scope based
on the presence of Numba, and these globals are referred to in the
instance checks.

Note that:

- Checks on function types are not modified, and Numba and Numba-CUDA
  are sharing types when co-installed.
- Some checks also accepted `None` - these have been slightly modified
  so that the assertion is only used when the variable is not `None`.
- Other checks, e.g. for `int` and `str` types, require no modification.