Always use `core` errors if `numba` is present (#634)
This PR fixes an issue where `numba-cuda` may raise a
`numba.cuda.core.errors.TypingError` in some situations where typing
fails, and a `numba.core.errors.TypingError` in others. In this PR we
take the approach of always using `numba.core.errors` in our codebase if
numba is present, else use only our own.

This PR also adds some missing cases of test coverage from the cuDF test
suite that exposed this issue when performing the numba upgrade. These
tests rely on catching the correct type of `TypingError` which was found
to vary.

---------

Co-authored-by: Graham Markall <535640+gmarkall@users.noreply.github.com>